+++ assets/js/components/GlobalSettings/UserInterfaceSettings.vue
@@ -95,13 +95,18 @@
 import TelemetrySettings from "../TelemetrySettings.vue";
 import FormRow from "../Helper/FormRow.vue";
 import SelectGroup from "../Helper/SelectGroup.vue";
-import { getLocalePreference, setLocalePreference, LOCALES, removeLocalePreference } from "@/i18n";
-import { getThemePreference, setThemePreference, THEMES } from "@/theme";
+import {
+	getLocalePreference,
+	setLocalePreference,
+	LOCALES,
+	removeLocalePreference,
+} from "@/i18n.ts";
+import { getThemePreference, setThemePreference } from "@/theme.ts";
 import { getUnits, setUnits, UNITS, is12hFormat, set12hFormat } from "@/units";
-import { getHiddenFeatures, setHiddenFeatures } from "@/featureflags";
+import { getHiddenFeatures, setHiddenFeatures } from "@/featureflags.ts";
 import { isApp } from "@/utils/native";
 import { defineComponent, type PropType } from "vue";
-import type { Sponsor } from "@/types/evcc";
+import { THEME, type Sponsor } from "@/types/evcc";
 
 const TIME_12H = "12";
 const TIME_24H = "24";
@@ -120,7 +125,7 @@ export default defineComponent({
 			timeFormat: is12hFormat() ? TIME_12H : TIME_24H,
 			hiddenFeatures: getHiddenFeatures(),
 			fullscreenActive: false,
-			THEMES,
+			THEMES: Object.values(THEME),
 			UNITS,
 			TIME_FORMATS: [TIME_24H, TIME_12H],
 		};
@@ -157,7 +162,8 @@ export default defineComponent({
 		},
 		language(value) {
 			const i18n = this.$root?.$i18n;
-			if (value) {
+			if (!i18n) return;
+			else if (value) {
 				setLocalePreference(i18n, value);
 			} else {
 				removeLocalePreference(i18n);
