+++ assets/js/components/Savings/Savings.vue
@@ -210,8 +210,8 @@ import LiveCommunity from "./LiveCommunity.vue";
 import TelemetrySettings from "../TelemetrySettings.vue";
 import CustomSelect from "../Helper/CustomSelect.vue";
 import settings from "@/settings";
-import api, { allowClientError } from "@/api";
-import { docsPrefix } from "@/i18n";
+import api, { allowClientError } from "@/api.ts";
+import { docsPrefix } from "@/i18n.ts";
 import { defineComponent, type PropType } from "vue";
 import type { CURRENCY, Rate, SelectOption, Sponsor as SponsorType } from "@/types/evcc";
 import type { Period } from "./types";
